T-SQL Swiss Knife Using the ScriptDOM Parser
Arvind Shyamsundar gave a talk on the T-SQL Script DOM parser to the SQL
PASS AppDev Virtual Chapter meeting on the 1st of November 2013. This
talk is available on YourTube at
https://www.youtube.com/watch?v=CciVxRFXgH8&index=1&list=WL. However, I
was unable to find the examples from that talk, and this is my attempt
to replicate these examples.

Any errors are mine and not Arvinds.